【AI】重构音频输出与效果器体系，新增多后端与GM2支持
- 引入 AudioSink 抽象，实现 ALSA/PulseAudio/Jack/PipeWire 四种实时后端，支持多采样深度与增益软限幅。
- 重构效果器 DSP 模块，实现完整的 XG 混响/合唱/失真/调制/均衡/动态类效果，以及 XG2.0 和声与串联合成效果。
- 用 DoubleBuffered 替换 RwLock 实现无锁参数快照，消除读写竞争与写者饥饿。
- 实现 GM2 通用 SysEx 映射（主音量/调音/效果参数/键位控制），补齐 GS 地址重映射。
- 新增 Active Sensing 心跳看门狗与多 MIDI 输入源（ALSA/Jack/PipeWire）。
- 优化内存布局，音色库与键位定义改为稀疏存储，大幅降低常驻内存占用。
- 补齐 XG 效果类型默认参数表，类型切换时自动加载对应参数，并通过集成测试验证渲染链路。

+++ src/voice_manager/sample_meta.rs
@@ -9,72 +9,72 @@ pub trait SampleMetaFactory<T, O> {
 
 #[derive(Debug)]
 pub struct SampleMeta {
-    // 省点内存吧，本来波形文件就很大了，直接用指针就得了
+    // Save some memory: the waveform files are already huge, so just use pointers
     pub pcm: Option<&'static [f32]>,
     pub loop_point: usize,
     pub loop_length: usize,
-    // 基准音高
+    // base pitch
     base_note: u8,
-    // 音高微调
+    // pitch fine-tune
     tone: i8,
     pub end_note: u8,
     pub sample_rate: u8,
     pub base_cent: f32,
 
-    /// 键位下限 (FUN_10017060: key_range 匹配)
+    /// Key lower bound (FUN_10017060: key_range matching)
     pub key_min: u8,
-    /// 键位上限
+    /// Key upper bound
     pub key_max: u8,
-    /// 力度下限
+    /// Velocity lower bound
     pub vel_min: u8,
-    /// 力度上限
+    /// Velocity upper bound
     pub vel_max: u8,
 
     // some parameters
-    // ── LFO/力度/音高/音量 (4 bytes) ══════════════════════════════
-    /// LFO 波形选择 (低3位 mask 0x7 = wave type); 高5位 = 鼓键号
+    // ── LFO/velocity/pitch/volume (4 bytes) ══════════════════════════════
+    /// LFO wave selection (low 3 bits mask 0x7 = wave type); high 5 bits = drum key number
     pub lfo_wave: u8,
-    /// 力度分层阈值 (FUN_10004DF0 读取)
+    /// Velocity layer threshold (read by FUN_10004DF0)
     pub vel_threshold: u8,
-    /// 音高偏移 (signed, -128~+127)
+    /// Pitch offset (signed, -128~+127)
     pub pitch_offset: i8,
-    /// 音量偏移 (signed, -128~+127)
+    /// Volume offset (signed, -128~+127)
     pub vol_offset: i8,
 
-    // ── 音高微调 (2 bytes) ════════════════════════════════════════
-    /// 音高微调: combined = (elem[9]-8)*256 + elem[10]*16 → 12-bit
+    // ── Pitch fine-tune (2 bytes) ════════════════════════════════════════
+    /// Pitch fine-tune: combined = (elem[9]-8)*256 + elem[10]*16 → 12-bit
     pub pitch_fine_h: i8,
     pub pitch_fine_l: i8,
 
     // ── Pitch EG + Filter (4 bytes) ═══════════════════════════════
-    /// Pitch EG Attack 速率 (0=最快)
+    /// Pitch EG Attack rate (0=fastest)
     pub pitch_eg_attack: u8,
-    /// Pitch EG Decay/Release 速率 (0=最快, 值域 0-3)
+    /// Pitch EG Decay/Release rate (0=fastest, range 0-3)
     pub pitch_eg_decay: u8,
-    /// 滤波器截止频率 (64=中心)
+    /// Filter cutoff frequency (64=center)
     pub filter_cutoff: u8,
-    /// 滤波器共鸣 (64=中心)
+    /// Filter resonance (64=center)
     pub filter_resonance: u8,
 
-    // ── 模式/范围/类型 (3 bytes) ══════════════════════════════════
-    /// 调音模式 (0=直加, 1-4=查表)
+    // ── Mode/range/type (3 bytes) ══════════════════════════════════
+    /// Pitch mode (0=direct add, 1-4=table lookup)
     pub pitch_mode: u8,
-    /// 范围基值 (60=中央C)
+    /// Range base value (60=middle C)
     pub range_base: u8,
-    /// 音色类型 (0=标准poly, 1=默认, 2=合成器, 3=SFX)
+    /// Voice type (0=standard poly, 1=default, 2=synthesizer, 3=SFX)
     pub voice_type: u8,
 
     // ── NoteShift + Detune + PEG front (4 bytes) ═════════════════
-    /// Note Shift (32-96, 64=中心); 运行时被覆写
+    /// Note Shift (32-96, 64=center); overwritten at runtime
     pub note_shift: u8,
-    /// Detune (14-114, 64=中心); 运行时被覆写
+    /// Detune (14-114, 64=center); overwritten at runtime
     pub detune: u8,
     /// PEG Center Note Low
     pub peg_center_low: u8,
     /// PEG Center Note High
     pub peg_center_high: u8,
 
-    // ── PEG 前段 (4 bytes) ═══════════════════════════════════════
+    // ── PEG front section (4 bytes) ═══════════════════════════════════════
     /// PEG Vel Sense Level
     pub peg_vel_sense_level: u8,
     /// PEG Vel Sense Rate
@@ -84,7 +84,7 @@ pub struct SampleMeta {
     /// PEG Center Note
     pub peg_center_note: u8,
 
-    // ── PEG 速率 (5 bytes, 值域 0-127) ═══════════════════════════
+    // ── PEG rates (5 bytes, range 0-127) ═══════════════════════════
     /// PEG Rate / Center (mode=64)
     pub peg_rate0: u8,
     /// PEG Rate 1
@@ -96,66 +96,66 @@ pub struct SampleMeta {
     /// PEG Rate 4
     pub peg_rate4: u8,
 
-    // ── DSP 合成参数 [31..77] ════════════════════════════════════
-    /// DSP 参数基索引 (FUN_10013456 读取 → vtable[0x4F0])
+    // ── DSP synthesis parameters [31..77] ════════════════════════════════════
+    /// DSP parameter base index (read by FUN_10013456 → vtable[0x4F0])
     pub dsp_base: u8,
-    /// 查表索引 → 音高缩放 (0x10015887 读取, 2D 查表 0x100473D0)
+    /// Table lookup index → pitch scaling (read by 0x10015887, 2D lookup 0x100473D0)
     pub tbl_index: u8,
-    /// 音高粗调偏移 (0x10006D4D 读取 → voice[0xB4])
+    /// Pitch coarse offset (read by 0x10006D4D → voice[0xB4])
     pub pitch_coarse: u8,
-    /// Filter EG 使能 (0x10019643: 非0→触发 Filter EG 链)
+    /// Filter EG enable (0x10019643: non-0 → triggers the Filter EG chain)
     pub eg_filt_en: u8,
-    /// Amp EG 使能 (0x10007315: 非0→计算 AEG → voice[0x6E])
+    /// Amp EG enable (0x10007315: non-0 → computes AEG → voice[0x6E])
     pub eg_amp_en: u8,
-    /// LFO 使能 (0x10007354)
+    /// LFO enable (0x10007354)
     pub lfo_en: u8,
-    /// Pitch EG 使能 (0x100073E6: 非0→调用 vtable[0x18C] PEG)
+    /// Pitch EG enable (0x100073E6: non-0 → calls vtable[0x18C] PEG)
     pub eg_pitch_en: u8,
-    /// 输出使能 (0x100074FA)
+    /// Output enable (0x100074FA)
     pub output_en: u8,
-    /// Cutoff 覆写标志 (0x10007834: 0=默认查表, 非0=覆写)
+    /// Cutoff override flag (0x10007834: 0=default table lookup, non-0=override)
     pub ovr_cutoff: u8,
-    /// Cutoff Scaling 阶段1 使能
+    /// Cutoff Scaling stage 1 enable
     pub cs_en_1: u8,
-    /// Cutoff Scaling 阶段2 使能
+    /// Cutoff Scaling stage 2 enable
     pub cs_en_2: u8,
-    /// Level Scaling 使能 (0x100196E1 → FUN_1001B630)
+    /// Level Scaling enable (0x100196E1 → FUN_1001B630)
     pub ls_en: u8,
-    /// Level Scaling 存储
+    /// Level Scaling storage
     pub ls_store: u8,
-    /// Level Scaling 后处理比较
+    /// Level Scaling post-processing comparison
     pub ls_cmp: u8,
-    /// Level Scaling 标志
+    /// Level Scaling flag
     pub ls_flag: u8,
-    /// AEG Decay1 Rate 覆写使能
+    /// AEG Decay1 Rate override enable
     pub aeg_d1: u8,
-    /// AEG Decay2 Rate 覆写使能
+    /// AEG Decay2 Rate override enable
     pub aeg_d2: u8,
-    /// AEG Release Rate 覆写使能
+    /// AEG Release Rate override enable
     pub aeg_rel: u8,
-    /// EG 速率重映射索引 (0x100142AC: 左移7→2D查表)
+    /// EG rate remap index (0x100142AC: shift left 7 → 2D table lookup)
     pub rate_idx: u8,
-    /// 采样格式标志 (0x10038D6C: 0=8bit, 非0=16bit)
+    /// Sample format flag (0x10038D6C: 0=8bit, non-0=16bit)
     pub fmt_flag: u8,
-    /// 查表索引 (0x10015834 → 字表 0x10048134 → voice[0xE])
+    /// Table lookup index (0x10015834 → word table 0x10048134 → voice[0xE])
     pub tbl_68: u8,
-    /// EG 阶段偏移/计数器 (0x10012550 → voice[0x1C7])
+    /// EG phase offset/counter (0x10012550 → voice[0x1C7])
     pub eg_phase: u8,
-    /// 采样层音高微调 (0x100125B0 → voice[0x1CA])
+    /// Sample-layer pitch fine-tune (0x100125B0 → voice[0x1CA])
     pub wave_pitch: u8,
-    /// EG 总使能 (0x10012600 → voice[0x64])
+    /// EG master enable (0x10012600 → voice[0x64])
     pub eg_enable: u8,
-    /// EG 延迟 (0x10012670 → voice[0x69][0x66])
+    /// EG delay (0x10012670 → voice[0x69][0x66])
     pub eg_delay: u8,
-    /// 触发/再触发模式 (0x100127A0)
+    /// Trigger/retrigger mode (0x100127A0)
     pub trig_mode: u8,
-    /// 覆写查表值 (0x10014238: 非0→覆写 0x10047AD8)
+    /// Override table lookup value (0x10014238: non-0 → overrides 0x10047AD8)
     pub alt_ovr: u8,
-    /// 采样偏移高位 (0x10015691 → voice[0x1CC])
+    /// Sample offset high bits (0x10015691 → voice[0x1CC])
     pub off_hi: u8,
-    /// 采样偏移低位 ([75] 低7位)
+    /// Sample offset low bits (low 7 bits of [75])
     pub off_lo: u8,
-    /// 灵敏度 signed (0x10013530: elem-64 → 调制 element[31])
+    /// Signed sensitivity (0x10013530: elem-64 → modulates element[31])
     pub sensitivity: u8,
 }
 
@@ -351,17 +351,17 @@ impl SampleMeta {
         ((self.pitch_coarse as i32 - 64) * 100) as f32
     }
 
-    /// 采样的基准音分（不含 tone）: base_note × 100
+    /// The sample's base pitch in cents (not including tone): base_note × 100
     pub fn get_base_note_cent(&self) -> f32 {
         self.base_note as f32 * 100.0
     }
 
-    /// 采样微调 (WaveEntry[2] tone, 音分)
+    /// Sample fine-tune (WaveEntry[2] tone, in cents)
     pub fn get_tone(&self) -> f32 {
         self.tone as f32
     }
 
-    /// element 音高偏移 (elem[7], signed, 音分)
+    /// Element pitch offset (elem[7], signed, in cents)
     pub fn get_pitch_offset(&self) -> f32 {
         self.pitch_offset as f32
     }
@@ -380,7 +380,7 @@ impl SampleMeta {
 
 #[inline(always)]
 fn to_cent(base_note: u8, tone: i8) -> f32 {
-    (base_note as i8 * 100 + tone) as f32
+    (base_note as i32 * 100 + tone as i32) as f32
 }
 
 pub const PITCH_FINE_TABLE_POS: [u16; 128] = [
